[surface] add acquire_snapshot_image
This patch adds another method that is similar to snapshot
but ensures that the returned surface is an image surface.
The method is useful in surface types for OS backends.
The patch for example uses it to eliminate unneeded copies
or replays in the Quartz backend, which was previously doing
acquire_source_image + snapshot.

The patch also provides implementations for the existing
surface types that already implemented the snapshot method.