[image/quartz] add image surface cache and use it for Quartz
The Quartz image was a special case of a more generic "cached
representation" concept usable by other non-vector surface types
(e.g. glitz or Win32).  This patch instead adds a cache residing in the
image surface, so that image surfaces now cache their representation
if requested by other surfaces.  The Quartz backend is correspondingly
updated to use the cache.  Cacheability is just a hint, and the user
should not assume any defined behavior if he/she modifies externally a
cacheable image surface and does not mark the surface as dirty.

The cache is a set of (rebuild-func, destroy-func, cached-value,
up-to-date) tuples indexed by an integer key (usually another surface
type).

The two internal functions to access it are _cairo_image_surface_get_cache
and _cairo_image_surface_set_cache.  The former can be used to lookup
a cached representation and optionally build it for the first time; the
latter can be used to discard a representation or set it.  The two
common use cases are calling _cairo_image_surface_get_cache with a
non-NULL rebuild-func, or calling _cairo_image_surface_get_cache with
a NULL rebuild-func and following it with _cairo_image_surface_set_cache
if the representation is not found.

The semantics of the rebuild functions are a bit hairy, but very
flexible.  The reconstruct function takes an image surface and returns
the cached representation that _cairo_image_surface_get_cache will
return.  If the entry is already present but dirty, the old value is
passed so that the underlying handle (e.g. a CGImageRef or HBITMAP)
can be recycled or freed.

Furthermore, the rebuild function can be called with NEEDED == true
(from _cairo_image_surface_get_cache) or NEEDED == false (when flushing).
Returning NULL just means that the rebuild function did not refresh the
representation and will cause _cairo_image_surface_get_cache to return
NULL as well; however, it will not remove the entry from the cache or
call the destroy function.  This allows to implement various policies
in the backends.

For example, if the rebuild is cheap (for example if the OS resource
can actually share the pixel data with the image), the rebuild
function can return a non-NULL value even when NEEDED is false.  On the
contrary, if the OS resources are scarce, the rebuild function can use
_cairo_image_surface_set_cache with NULL data when NEEDED is false to
discard the representation and have it rebuilt from scratch later.
In the normal case, the backend will return NULL when NEEDED is false.

im/quartz