[glitz] use image cache
This patch changes the Glitz backend to use the image cache.

I tested this patch on the cairogears benchmark.  I removed
the code to copy the PNG files from image surfaces to Glitz
surfaces, which caused performance to fall by up to 7-10x
(for example I had 40 fps instead of 360 on the SHADOW
test).  Making the images cacheable brought back performance
to those of the vanilla benchmark, while removing almost 100
lines of code from the benchmark.