[quartz] fix bug 18632, crash with gradients if the pattern is stack-allocated
The lifetime of the Quartz patterns for the source (created by
_cairo_quartz_setup_source and destroyed by _cairo_quartz_teardown_source)
always coincides within the execution of _cairo_quartz_setup_source's
caller.

Therefore, there is no need to reference-count the patterns passed
to _cairo_quartz_setup_source, and this is in fact wrong when
cairo-gstate.c is passing a stack-allocated pattern.